MDL-77576 core_communication: Add room membership api
This commit implements the room membership api for communication
in order to add users to a communication provider room when
a new instance is created, updated or the api is called to action
any member related action.

Originally implemented as MDL-76703.

+++ communication/classes/communication.php
@@ -41,16 +41,34 @@ class communication {
     public communication_settings_data $communicationsettings;
 
     /**
+     * @var communication_user_base $communicationuser The communication user object
+     */
+    public communication_user_base $communicationuser;
+
+    /**
+     * @var array $userids The id of the users
+     */
+    protected array $userids;
+
+    /**
      * Communication room constructor to get the communication features.
      *
      * @param int $instanceid The id of the instance
      * @param string $component The component of the instance
      * @param string $instancetype The type of instance for the component
      * @param string|null $instanceavatarurl The url of the avatar for the instance
+     * @param string|null $disableprovider The provider to use for associated tasks after disabled
+     * @param array $userids The user ids
      */
-    public function __construct(int $instanceid, string $component, string $instancetype, string $instanceavatarurl = null) {
+    public function __construct(int $instanceid, string $component, string $instancetype, string $instanceavatarurl = null,
+        string $disableprovider = null, array $userids = []) {
         $this->instanceavatarurl = $instanceavatarurl;
+        $this->userids = $userids;
         $this->communicationsettings = new communication_settings_data($instanceid, $component, $instancetype);
+        if ($disableprovider !== null) {
+            $this->communicationsettings->provider = $disableprovider;
+            $this->communicationsettings->disableprovider = $disableprovider;
+        }
         $this->init_provider();
     }
 
@@ -77,6 +95,12 @@ class communication {
             if (!empty($communicationroom)) {
                 $this->communicationroom = $communicationroom;
             }
+            if (!empty($this->userids)) {
+                $communicationuser = $pluginentrypoint->get_provider_user($this);
+                if (!empty($communicationuser)) {
+                    $this->communicationuser = $communicationuser;
+                }
+            }
         }
     }
 
@@ -98,7 +122,8 @@ class communication {
      * @return void
      */
     public function create_room(): void {
-        if ($this->check_object_and_method_exist('communicationroom', 'create')) {
+        if ($this->check_object_and_method_exist('communicationroom', 'create') &&
+                $this->communicationsettings->disableprovider === null) {
             $this->communicationroom->create();
         }
     }
@@ -109,7 +134,8 @@ class communication {
      * @return void
      */
     public function update_room(): void {
-        if ($this->check_object_and_method_exist('communicationroom', 'update')) {
+        if ($this->check_object_and_method_exist('communicationroom', 'update') &&
+                $this->communicationsettings->disableprovider === null) {
             $this->communicationroom->update();
         }
     }
@@ -120,11 +146,13 @@ class communication {
      * @return void
      */
     public function delete_room(): void {
-        if ($this->check_object_and_method_exist('communicationroom', 'delete')) {
-            $this->communicationroom->delete();
+        if ($this->communicationsettings->disableprovider === null) {
+            if ($this->check_object_and_method_exist('communicationroom', 'delete')) {
+                $this->communicationroom->delete();
+            }
+            // Now delete the local communication record after the deletion if done from the plugin.
+            $this->communicationsettings->delete();
         }
-        // Now delete the local communication record after the deletion if done from the plugin.
-        $this->communicationsettings->delete();
     }
 
     /**
@@ -137,5 +165,27 @@ class communication {
             return $this->communicationroom->generate_room_url();
         }
         return null;
+    }
+
+    /**
+     * Add members to the room.
+     *
+     * @return void
+     */
+    public function add_members(): void {
+        if ($this->check_object_and_method_exist('communicationuser', 'add_members_to_room')) {
+            $this->communicationuser->add_members_to_room($this->userids);
+        }
+    }
+
+    /**
+     * Remove members from room.
+     *
+     * @return void
+     */
+    public function remove_members(): void {
+        if ($this->check_object_and_method_exist('communicationuser', 'remove_members_from_room')) {
+            $this->communicationuser->remove_members_from_room($this->userids);
+        }
     }
 }
